fix: r1-fix — distribute tests inline per telegram sub-module
49 tests distributed to their locality sub-modules:
- state.rs: 8 tests (320 LOC total)
- topic_registry.rs: 17 tests (375 LOC total)
- send.rs: 13 tests (302 LOC total)
- inbound.rs: 1 test (499 LOC total)
- error.rs: 10 tests (346 LOC total)
- mod.rs: 26 PR-3b tests remain (transition state)

All sub-files ≤499 LOC (well under 700).
