+++ src/channel/telegram/error.rs
@@ -149,3 +149,198 @@ pub(crate) fn invalidate_and_recreate_topic(
     );
     create_topic_for_instance(home, instance_name)
 }
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used)]
+mod tests {
+    use super::*;
+    use std::collections::HashMap;
+    use std::path::PathBuf;
+
+    use crate::channel::telegram::state::TelegramState;
+    use crate::channel::telegram::topic_registry::{
+        load_topic_registry, register_topic, save_topic_registry,
+    };
+
+    fn tmp_home(name: &str) -> PathBuf {
+        use std::sync::atomic::{AtomicU32, Ordering};
+        static COUNTER: AtomicU32 = AtomicU32::new(0);
+        let id = COUNTER.fetch_add(1, Ordering::Relaxed);
+        let dir = std::env::temp_dir().join(format!(
+            "agend-telegram-test-{}-{}-{}",
+            std::process::id(),
+            name,
+            id
+        ));
+        std::fs::create_dir_all(&dir).ok();
+        dir
+    }
+
+    #[test]
+    fn is_topic_deleted_error_matches_thread_not_found() {
+        let e = anyhow::anyhow!("Bad Request: message thread not found");
+        assert!(is_topic_deleted_error(&e));
+    }
+
+    #[test]
+    fn is_topic_deleted_error_case_insensitive() {
+        let e = anyhow::anyhow!("BAD REQUEST: MESSAGE THREAD NOT FOUND");
+        assert!(is_topic_deleted_error(&e));
+    }
+
+    #[test]
+    fn is_topic_deleted_error_matches_wrapped_context() {
+        let inner = anyhow::anyhow!("Bad Request: message thread not found");
+        let wrapped = inner.context("sending to topic 42");
+        assert!(is_topic_deleted_error(&wrapped));
+    }
+
+    #[test]
+    fn is_topic_deleted_error_rejects_unrelated() {
+        for msg in [
+            "network timeout",
+            "Bad Request: chat not found",
+            "Forbidden: bot was blocked by the user",
+            "Too Many Requests: retry after 5",
+            "Bad Request: message to edit not found",
+        ] {
+            let e = anyhow::anyhow!(msg.to_string());
+            assert!(
+                !is_topic_deleted_error(&e),
+                "classifier must not match: {msg}"
+            );
+        }
+    }
+
+    #[test]
+    fn cleanup_deleted_topic_clears_state_maps() {
+        let home = tmp_home("cleanup-state");
+        let mut topic_map = HashMap::new();
+        topic_map.insert("agent1".to_string(), 42);
+        topic_map.insert("agent2".to_string(), 43);
+        let mut submit_keys = HashMap::new();
+        submit_keys.insert("agent1".to_string(), "\r".to_string());
+        let state = Arc::new(Mutex::new(TelegramState::new(
+            "tok",
+            -1,
+            topic_map,
+            home.clone(),
+            submit_keys,
+            None,
+        )));
+
+        cleanup_deleted_topic(&home, "agent1", 42, Some(&state));
+
+        let s = state.lock();
+        assert!(!s.topic_to_instance.contains_key(&42));
+        assert!(!s.instance_to_topic.contains_key("agent1"));
+        assert!(!s.submit_keys.contains_key("agent1"));
+        assert_eq!(s.topic_to_instance.get(&43), Some(&"agent2".to_string()));
+        assert_eq!(s.instance_to_topic.get("agent2"), Some(&43));
+    }
+
+    #[test]
+    fn cleanup_deleted_topic_unregisters_topic() {
+        let home = tmp_home("cleanup-registry");
+        let mut reg = HashMap::new();
+        reg.insert(99, "ghost".to_string());
+        reg.insert(100, "alive".to_string());
+        save_topic_registry(&home, &reg);
+
+        cleanup_deleted_topic(&home, "ghost", 99, None);
+
+        let after = load_topic_registry(&home);
+        assert!(!after.contains_key(&99));
+        assert_eq!(after.get(&100), Some(&"alive".to_string()));
+    }
+
+    #[test]
+    fn handle_send_failure_only_fires_on_topic_deleted() {
+        let home = tmp_home("handle-send");
+        let unrelated = anyhow::anyhow!("network timeout");
+        assert!(!handle_send_failure(
+            &unrelated,
+            &home,
+            "any",
+            Some(42),
+            None
+        ));
+        let gone = anyhow::anyhow!("Bad Request: message thread not found");
+        assert!(!handle_send_failure(&gone, &home, "any", None, None));
+        assert!(handle_send_failure(&gone, &home, "any", Some(42), None));
+    }
+
+    #[test]
+    fn invalidate_and_recreate_strips_stale_topic_from_registry() {
+        let home = tmp_home("invalidate-recreate-strip");
+        register_topic(&home, 42, "agent-x");
+        register_topic(&home, 99, "agent-y");
+
+        let result = invalidate_and_recreate_topic(&home, "agent-x", 42);
+        assert!(result.is_none(), "no bot → creation fails");
+
+        let reg = load_topic_registry(&home);
+        assert!(
+            !reg.contains_key(&42),
+            "stale topic 42 must be unregistered"
+        );
+        assert_eq!(reg.get(&99), Some(&"agent-y".to_string()));
+        std::fs::remove_dir_all(&home).ok();
+    }
+
+    #[test]
+    fn invalidate_and_recreate_clears_fleet_yaml_topic_id() {
+        let home = tmp_home("invalidate-recreate-yaml");
+        let yaml = "\
+instances:
+  agent-x:
+    command: /bin/true
+    topic_id: 42
+  agent-y:
+    command: /bin/true
+    topic_id: 99
+";
+        std::fs::write(home.join("fleet.yaml"), yaml).expect("write fleet.yaml");
+        register_topic(&home, 42, "agent-x");
+
+        let _ = invalidate_and_recreate_topic(&home, "agent-x", 42);
+
+        let config =
+            crate::fleet::FleetConfig::load(&home.join("fleet.yaml")).expect("load fleet.yaml");
+        let agent_x = config.instances.get("agent-x").expect("agent-x exists");
+        assert_eq!(
+            agent_x.topic_id, None,
+            "topic_id must be cleared after invalidation"
+        );
+        let agent_y = config.instances.get("agent-y").expect("agent-y exists");
+        assert_eq!(agent_y.topic_id, Some(99));
+        std::fs::remove_dir_all(&home).ok();
+    }
+
+    #[test]
+    fn invalidate_and_recreate_preserves_instance_in_fleet_yaml() {
+        let home = tmp_home("invalidate-preserves-instance");
+        let yaml = "\
+instances:
+  agent-x:
+    command: /bin/true
+    topic_id: 42
+    role: important
+";
+        std::fs::write(home.join("fleet.yaml"), yaml).expect("write fleet.yaml");
+        register_topic(&home, 42, "agent-x");
+
+        let _ = invalidate_and_recreate_topic(&home, "agent-x", 42);
+
+        let fleet_yaml = std::fs::read_to_string(home.join("fleet.yaml")).expect("read fleet.yaml");
+        assert!(
+            fleet_yaml.contains("agent-x"),
+            "instance must survive invalidation; yaml:\n{fleet_yaml}"
+        );
+        assert!(
+            fleet_yaml.contains("important"),
+            "instance role must survive; yaml:\n{fleet_yaml}"
+        );
+        std::fs::remove_dir_all(&home).ok();
+    }
+}
