+++ src/channel/telegram/mod.rs
@@ -1422,169 +1422,6 @@ mod tests {
     }
 
     #[test]
-    fn telegram_state_new_builds_reverse_map() {
-        let mut topic_map = HashMap::new();
-        topic_map.insert("agent1".to_string(), 100);
-        topic_map.insert("agent2".to_string(), 200);
-        let state = TelegramState::new(
-            "fake-token",
-            -12345,
-            topic_map,
-            PathBuf::from("/tmp/test"),
-            HashMap::new(),
-            None,
-        );
-        assert_eq!(
-            state.topic_to_instance.get(&100),
-            Some(&"agent1".to_string())
-        );
-        assert_eq!(
-            state.topic_to_instance.get(&200),
-            Some(&"agent2".to_string())
-        );
-        assert_eq!(state.instance_to_topic.get("agent1"), Some(&100));
-        assert_eq!(state.instance_to_topic.get("agent2"), Some(&200));
-    }
-
-    #[test]
-    fn telegram_state_empty_topic_map() {
-        let state = TelegramState::new(
-            "fake-token",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        assert!(state.topic_to_instance.is_empty());
-        assert!(state.instance_to_topic.is_empty());
-    }
-
-    #[test]
-    fn telegram_state_submit_keys_preserved() {
-        let mut keys = HashMap::new();
-        keys.insert("agent1".to_string(), "\n".to_string());
-        let state =
-            TelegramState::new("tok", -1, HashMap::new(), PathBuf::from("/tmp"), keys, None);
-        assert_eq!(state.submit_keys.get("agent1"), Some(&"\n".to_string()));
-    }
-
-    #[test]
-    fn is_user_allowed_none_rejects_all_after_phase2_fail_closed_swap() {
-        // Sprint 21 Phase 2 inversion: previously `None` allowlist
-        // accepted all (legacy fail-open). Now `None` rejects all
-        // (fail-closed) — operator must configure `user_allowlist:
-        // [user_id, ...]` to authorise senders. See PR #216 + Phase 2
-        // dispatch (cascade auth chain inbound side).
-        let state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        assert!(
-            !state.is_user_allowed(1),
-            "fail-closed: None must reject any user"
-        );
-        assert!(
-            !state.is_user_allowed(i64::MAX),
-            "fail-closed: None must reject even valid Telegram user_ids"
-        );
-    }
-
-    #[test]
-    fn is_user_allowed_empty_rejects_all() {
-        let state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            Some(vec![]),
-        );
-        assert!(!state.is_user_allowed(1));
-        assert!(!state.is_user_allowed(0));
-    }
-
-    #[test]
-    fn is_user_allowed_restricts_to_list() {
-        let state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            Some(vec![42, 100]),
-        );
-        assert!(state.is_user_allowed(42));
-        assert!(state.is_user_allowed(100));
-        assert!(!state.is_user_allowed(41));
-        assert!(!state.is_user_allowed(0));
-    }
-
-    #[test]
-    fn is_user_allowed_realistic_user_id_after_phase2() {
-        // Operator-pitfall regression: real Telegram user_ids are i64
-        // values in the 100-million-to-billion range (e.g.
-        // `1047180393`). YAML deserialisation via `serde_yaml_ng` must
-        // round-trip these as i64, not String — and the predicate
-        // must compare i64 equality, not string equality. This pin
-        // catches any future schema drift that lossy-coerces user_id.
-        const REAL_USER_ID: i64 = 1_047_180_393;
-        let state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            Some(vec![REAL_USER_ID]),
-        );
-        assert!(state.is_user_allowed(REAL_USER_ID));
-        // Off-by-one user_id is rejected — confirms exact i64 compare.
-        assert!(!state.is_user_allowed(REAL_USER_ID + 1));
-        assert!(!state.is_user_allowed(REAL_USER_ID - 1));
-    }
-
-    #[test]
-    fn user_allowlist_yaml_round_trip_i64_realistic_values() {
-        // Operator-pitfall regression (per Phase 2 dispatch): YAML
-        // round-trip for both `user_allowlist: [<positive i64>]` and
-        // `group_id: <negative supergroup i64>` must preserve i64
-        // semantics. Without this, a future serde refactor that
-        // narrowed to i32 would silently truncate user_ids beyond 2^31.
-        //
-        // The values used are the real-world shape operator confirmed
-        // via telegram (positive 10-digit user_id; negative 13-digit
-        // supergroup chat id with `-100` prefix per Telegram Bot API).
-        let yaml = r#"
-group_id: -1003725098111
-user_allowlist:
-  - 1047180393
-"#;
-        #[derive(serde::Deserialize, Debug)]
-        struct PartialChannel {
-            group_id: i64,
-            user_allowlist: Vec<i64>,
-        }
-        let parsed: PartialChannel = serde_yaml_ng::from_str(yaml).expect("yaml must deserialize");
-        assert_eq!(parsed.group_id, -1_003_725_098_111_i64);
-        assert_eq!(parsed.user_allowlist, vec![1_047_180_393_i64]);
-
-        // Round-trip back to YAML and re-parse to lock contract.
-        let serialized = serde_yaml_ng::to_string(&serde_json::json!({
-            "group_id": parsed.group_id,
-            "user_allowlist": parsed.user_allowlist,
-        }))
-        .expect("yaml must serialize");
-        let reparsed: PartialChannel =
-            serde_yaml_ng::from_str(&serialized).expect("yaml round-trip must deserialize");
-        assert_eq!(reparsed.group_id, parsed.group_id);
-        assert_eq!(reparsed.user_allowlist, parsed.user_allowlist);
-    }
-
-    #[test]
     fn map_emoji_name_known() {
         assert_eq!(map_emoji_name("thumbsup"), "👍");
         assert_eq!(map_emoji_name("thumbs_up"), "👍");
@@ -1946,359 +1783,6 @@ user_allowlist:
         ));
         std::fs::create_dir_all(&dir).ok();
         dir
-    }
-
-    #[test]
-    fn resolve_topic_known_topic() {
-        let mut topic_map = HashMap::new();
-        topic_map.insert("alice".to_string(), 100);
-        let mut state = TelegramState::new(
-            "tok",
-            -1,
-            topic_map,
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        assert_eq!(resolve_topic(&mut state, Some(100)), "alice");
-    }
-
-    #[test]
-    fn resolve_topic_none_returns_general() {
-        let mut state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        assert_eq!(resolve_topic(&mut state, None), "general");
-    }
-
-    #[test]
-    fn resolve_topic_unknown_falls_back_to_general() {
-        let mut state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp/nonexistent"),
-            HashMap::new(),
-            None,
-        );
-        // No fleet.yaml → falls back to general
-        assert_eq!(resolve_topic(&mut state, Some(999)), "general");
-    }
-
-    #[test]
-    fn resolve_topic_reloads_from_fleet_yaml() {
-        let home = tmp_home("resolve_reload");
-        let yaml = r#"defaults:
-  backend: claude
-instances:
-  alice:
-    role: "Test"
-    topic_id: 229
-  general:
-    role: "General"
-    topic_id: 1
-"#;
-        std::fs::write(home.join("fleet.yaml"), yaml).ok();
-
-        // State has NO topic mappings — simulates runtime-created topic
-        let mut state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            home.clone(),
-            HashMap::new(),
-            None,
-        );
-
-        // Should reload from fleet.yaml and find alice
-        assert_eq!(resolve_topic(&mut state, Some(229)), "alice");
-        // Should be cached now
-        assert_eq!(
-            state.topic_to_instance.get(&229),
-            Some(&"alice".to_string())
-        );
-        assert_eq!(state.instance_to_topic.get("alice"), Some(&229));
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn resolve_topic_reload_caches_for_next_call() {
-        let home = tmp_home("resolve_cache");
-        let yaml = r#"instances:
-  bob:
-    topic_id: 500
-"#;
-        std::fs::write(home.join("fleet.yaml"), yaml).ok();
-        let mut state = TelegramState::new(
-            "tok",
-            -1,
-            HashMap::new(),
-            home.clone(),
-            HashMap::new(),
-            None,
-        );
-
-        // First call: reloads from fleet.yaml
-        assert_eq!(resolve_topic(&mut state, Some(500)), "bob");
-        // Delete fleet.yaml — second call should use cached map
-        std::fs::remove_file(home.join("fleet.yaml")).ok();
-        assert_eq!(resolve_topic(&mut state, Some(500)), "bob");
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// Fix B: resolve_topic checks topics.json as third source when
-    /// topic_id is not in memory or fleet.yaml. Before this fix,
-    /// runtime-created instances (registered only in topics.json)
-    /// would misroute to "general".
-    #[test]
-    fn resolve_topic_falls_back_to_topics_json() {
-        let home = tmp_home("resolve_topics_json");
-        // No fleet.yaml — topic only exists in topics.json.
-        // Use save_topic_registry directly (not register_topic, which
-        // now writes fleet.yaml too — we want to test the read fallback).
-        let mut reg = HashMap::new();
-        reg.insert(2474, "test-gemini".to_string());
-        save_topic_registry(&home, &reg);
-        let mut state = TelegramState::new_for_contract_test(
-            -1,
-            HashMap::new(),
-            home.clone(),
-            HashMap::new(),
-            None,
-        );
-        assert_eq!(
-            resolve_topic(&mut state, Some(2474)),
-            "test-gemini",
-            "topic in topics.json must resolve, not fall through to general"
-        );
-        assert_eq!(resolve_topic(&mut state, Some(2474)), "test-gemini");
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// Fix B write-side: register_topic writes fleet.yaml topic_id.
-    /// §3.5.11: on main HEAD (before fix), register_topic does NOT
-    /// write fleet.yaml — this test would fail.
-    #[test]
-    fn register_topic_writes_fleet_yaml() {
-        let home = tmp_home("register_writes_yaml");
-        // Create a minimal fleet.yaml with the instance but no topic_id.
-        let yaml = "defaults:\n  backend: claude\ninstances:\n  alice:\n    backend: claude\n";
-        std::fs::write(home.join("fleet.yaml"), yaml).ok();
-
-        register_topic(&home, 42, "alice");
-
-        // Verify fleet.yaml now contains topic_id: 42 for alice.
-        let content =
-            std::fs::read_to_string(home.join("fleet.yaml")).expect("fleet.yaml must exist");
-        assert!(
-            content.contains("topic_id"),
-            "fleet.yaml must contain topic_id after register_topic: {content}"
-        );
-        // Also verify topics.json was written.
-        let reg = load_topic_registry(&home);
-        assert_eq!(reg.get(&42), Some(&"alice".to_string()));
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    // --- Topic registry tests ---
-
-    #[test]
-    fn topic_registry_roundtrip() {
-        let home = tmp_home("registry_roundtrip");
-        assert!(load_topic_registry(&home).is_empty());
-
-        register_topic(&home, 100, "alice");
-        register_topic(&home, 200, "bob");
-
-        let reg = load_topic_registry(&home);
-        assert_eq!(reg.get(&100), Some(&"alice".to_string()));
-        assert_eq!(reg.get(&200), Some(&"bob".to_string()));
-
-        unregister_topic(&home, 100);
-        let reg = load_topic_registry(&home);
-        assert!(!reg.contains_key(&100));
-        assert_eq!(reg.get(&200), Some(&"bob".to_string()));
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn topic_registry_overwrite() {
-        let home = tmp_home("registry_overwrite");
-        register_topic(&home, 100, "alice");
-        register_topic(&home, 100, "bob");
-
-        let reg = load_topic_registry(&home);
-        assert_eq!(reg.get(&100), Some(&"bob".to_string()));
-        assert_eq!(reg.len(), 1);
-
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn topic_registry_missing_file() {
-        let home = PathBuf::from("/tmp/agend-test-nonexistent-dir-12345");
-        let reg = load_topic_registry(&home);
-        assert!(reg.is_empty());
-    }
-
-    /// Regression pin: create_topic_for_instance must resolve credentials
-    /// from the caller-supplied home, not from `AGEND_HOME` / real
-    /// `home_dir()`. Before this fix the helper called
-    /// `resolve_channel_only()` which re-read the operator's real
-    /// fleet.yaml, so `cargo test` runs of the `positive_pin`
-    /// dispatch test (src/api/mod.rs) leaked into the live Telegram
-    /// group as a stray `positive_pin-1` topic. Here the test home has
-    /// no fleet.yaml at all; the helper must see "no channel
-    /// configured" and return `None` silently — no network call, no
-    /// leak.
-    #[test]
-    fn create_topic_for_instance_uses_passed_home_not_real_home() {
-        let home = tmp_home("topic-helper-home-scope");
-        // Deliberately no fleet.yaml → resolve_channel_only_from must
-        // fail at the fleet.yaml load step. Even if AGEND_BOT_TOKEN is
-        // set in the environment, we must not reach the teloxide call
-        // because the test home has no channel config to resolve
-        // group_id from.
-        let result = create_topic_for_instance(&home, "regression-pin");
-        assert!(
-            result.is_none(),
-            "missing fleet.yaml in the passed home must suppress the API call, got {result:?}"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// Companion to the test above for the symmetric delete path.
-    /// `delete_topic` also accepts `home` but used the unscoped resolver;
-    /// in tests that exercise teardown against a temp home it would have
-    /// attempted to hit the real Telegram group on a made-up topic id.
-    #[test]
-    fn delete_topic_uses_passed_home_not_real_home() {
-        let home = tmp_home("delete-topic-home-scope");
-        // No fleet.yaml → should return without touching the network.
-        // Can't directly assert "no HTTP request", but the early-return
-        // branch is the only code path that terminates cleanly when
-        // resolve fails, so reaching this assertion at all is proof.
-        delete_topic(&home, 999_999);
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn is_topic_deleted_error_matches_thread_not_found() {
-        // Exact wording returned by Telegram Bot API when the thread was deleted.
-        let e = anyhow::anyhow!("Bad Request: message thread not found");
-        assert!(is_topic_deleted_error(&e));
-    }
-
-    #[test]
-    fn is_topic_deleted_error_case_insensitive() {
-        let e = anyhow::anyhow!("BAD REQUEST: MESSAGE THREAD NOT FOUND");
-        assert!(is_topic_deleted_error(&e));
-    }
-
-    #[test]
-    fn is_topic_deleted_error_matches_wrapped_context() {
-        // Errors often get `.context(...)` wrapped; Alternate debug formatter
-        // (`{:#}`) flattens the chain into a single string we can inspect.
-        let inner = anyhow::anyhow!("Bad Request: message thread not found");
-        let wrapped = inner.context("sending to topic 42");
-        assert!(is_topic_deleted_error(&wrapped));
-    }
-
-    #[test]
-    fn is_topic_deleted_error_rejects_unrelated() {
-        // Common transient / auth errors must NOT trigger cleanup.
-        for msg in [
-            "network timeout",
-            "Bad Request: chat not found",
-            "Forbidden: bot was blocked by the user",
-            "Too Many Requests: retry after 5",
-            "Bad Request: message to edit not found",
-        ] {
-            let e = anyhow::anyhow!(msg.to_string());
-            assert!(
-                !is_topic_deleted_error(&e),
-                "classifier must not match: {msg}"
-            );
-        }
-    }
-
-    #[test]
-    fn cleanup_deleted_topic_clears_state_maps() {
-        // State-aware cleanup path: topic_to_instance / instance_to_topic /
-        // submit_keys entries must all be purged atomically. api::call and
-        // fleet.yaml removal fail silently when there's no daemon/file — that
-        // matches the production contract (we log and move on).
-        use parking_lot::Mutex;
-        use std::sync::Arc;
-        let home = tmp_home("cleanup-state");
-        let mut topic_map = HashMap::new();
-        topic_map.insert("agent1".to_string(), 42);
-        topic_map.insert("agent2".to_string(), 43);
-        let mut submit_keys = HashMap::new();
-        submit_keys.insert("agent1".to_string(), "\r".to_string());
-        let state = Arc::new(Mutex::new(TelegramState::new(
-            "tok",
-            -1,
-            topic_map,
-            home.clone(),
-            submit_keys,
-            None,
-        )));
-
-        cleanup_deleted_topic(&home, "agent1", 42, Some(&state));
-
-        let s = state.lock();
-        assert!(!s.topic_to_instance.contains_key(&42));
-        assert!(!s.instance_to_topic.contains_key("agent1"));
-        assert!(!s.submit_keys.contains_key("agent1"));
-        // agent2 untouched.
-        assert_eq!(s.topic_to_instance.get(&43), Some(&"agent2".to_string()));
-        assert_eq!(s.instance_to_topic.get("agent2"), Some(&43));
-    }
-
-    #[test]
-    fn cleanup_deleted_topic_unregisters_topic() {
-        // Both state-aware and headless callers must strip the topic from the
-        // on-disk registry so orphan-cleanup on next startup doesn't re-see it.
-        let home = tmp_home("cleanup-registry");
-        let mut reg = HashMap::new();
-        reg.insert(99, "ghost".to_string());
-        reg.insert(100, "alive".to_string());
-        save_topic_registry(&home, &reg);
-
-        cleanup_deleted_topic(&home, "ghost", 99, None);
-
-        let after = load_topic_registry(&home);
-        assert!(!after.contains_key(&99));
-        assert_eq!(after.get(&100), Some(&"alive".to_string()));
-    }
-
-    #[test]
-    fn handle_send_failure_only_fires_on_topic_deleted() {
-        let home = tmp_home("handle-send");
-        // Unrelated error: classifier miss → returns false, no cleanup.
-        let unrelated = anyhow::anyhow!("network timeout");
-        assert!(!handle_send_failure(
-            &unrelated,
-            &home,
-            "any",
-            Some(42),
-            None
-        ));
-        // Topic-deleted error but no topic_id → nothing to clean up.
-        let gone = anyhow::anyhow!("Bad Request: message thread not found");
-        assert!(!handle_send_failure(&gone, &home, "any", None, None));
-        // Topic-deleted + topic_id → fires.
-        assert!(handle_send_failure(&gone, &home, "any", Some(42), None));
     }
 
     /// Value-source pin (Reviewer Contract v0.1 §4) for S2d provenance.
@@ -2776,313 +2260,6 @@ instances:
                 "error must name the missing config for {label}: {err_str}",
             );
         }
-    }
-
-    // ─── Outbound media helper tests (Phase 3 PR-AG) ──────────────────
-
-    fn make_attachment(kind: crate::channel::AttachmentKind) -> crate::channel::Attachment {
-        crate::channel::Attachment {
-            kind,
-            path: PathBuf::from("/tmp/test.jpg"),
-            mime: None,
-            caption: None,
-            size_bytes: None,
-            original_filename: None,
-        }
-    }
-
-    #[test]
-    fn resolve_caption_uses_text_when_short() {
-        let att = make_attachment(crate::channel::AttachmentKind::Photo);
-        assert_eq!(resolve_caption("hello", &att).as_deref(), Some("hello"));
-    }
-
-    #[test]
-    fn resolve_caption_none_when_text_exceeds_limit() {
-        let att = make_attachment(crate::channel::AttachmentKind::Photo);
-        assert!(resolve_caption(&"x".repeat(CAPTION_MAX_CHARS + 1), &att).is_none());
-    }
-
-    #[test]
-    fn resolve_caption_prefers_attachment_caption() {
-        let mut att = make_attachment(crate::channel::AttachmentKind::Document);
-        att.caption = Some("explicit".into());
-        assert_eq!(resolve_caption("text", &att).as_deref(), Some("explicit"));
-    }
-
-    #[test]
-    fn resolve_caption_truncates_long_attachment_caption() {
-        let mut att = make_attachment(crate::channel::AttachmentKind::Photo);
-        att.caption = Some("x".repeat(CAPTION_MAX_CHARS + 100));
-        let cap = resolve_caption("", &att).expect("should have caption");
-        assert_eq!(cap.chars().count(), CAPTION_MAX_CHARS);
-    }
-
-    #[test]
-    fn resolve_caption_none_for_sticker() {
-        let att = make_attachment(crate::channel::AttachmentKind::Sticker);
-        assert!(resolve_caption("hello", &att).is_none());
-    }
-
-    #[test]
-    fn resolve_caption_none_when_text_empty() {
-        let att = make_attachment(crate::channel::AttachmentKind::Photo);
-        assert!(resolve_caption("", &att).is_none());
-    }
-
-    #[test]
-    fn needs_separate_text_false_when_empty() {
-        assert!(!needs_separate_text(
-            "",
-            &make_attachment(crate::channel::AttachmentKind::Photo)
-        ));
-    }
-
-    #[test]
-    fn needs_separate_text_false_when_short() {
-        assert!(!needs_separate_text(
-            "short",
-            &make_attachment(crate::channel::AttachmentKind::Photo)
-        ));
-    }
-
-    #[test]
-    fn needs_separate_text_true_when_long() {
-        assert!(needs_separate_text(
-            &"x".repeat(CAPTION_MAX_CHARS + 1),
-            &make_attachment(crate::channel::AttachmentKind::Photo)
-        ));
-    }
-
-    #[test]
-    fn needs_separate_text_true_for_sticker_with_text() {
-        assert!(needs_separate_text(
-            "hello",
-            &make_attachment(crate::channel::AttachmentKind::Sticker)
-        ));
-    }
-
-    #[test]
-    fn needs_separate_text_true_when_attachment_has_own_caption() {
-        let mut att = make_attachment(crate::channel::AttachmentKind::Photo);
-        att.caption = Some("cap".into());
-        assert!(needs_separate_text("body", &att));
-    }
-
-    #[test]
-    fn send_text_only_reaches_text_path() {
-        use crate::channel::Channel;
-        let state = TelegramState::new_for_contract_test(
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        let channel = TelegramChannel::new(Arc::new(Mutex::new(state)));
-        let binding = crate::channel::BindingRef::new("telegram", None, ());
-        let err = channel
-            .send(&binding, crate::channel::OutMsg::text("hello"))
-            .expect_err("no bot");
-        assert!(
-            err.to_string().contains("telegram bot not initialized"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn send_attachment_reaches_media_path() {
-        use crate::channel::Channel;
-        let state = TelegramState::new_for_contract_test(
-            -1,
-            HashMap::new(),
-            PathBuf::from("/tmp"),
-            HashMap::new(),
-            None,
-        );
-        let channel = TelegramChannel::new(Arc::new(Mutex::new(state)));
-        let binding = crate::channel::BindingRef::new("telegram", None, ());
-        let msg = crate::channel::OutMsg {
-            text: "see".into(),
-            attachment: Some(make_attachment(crate::channel::AttachmentKind::Photo)),
-            in_reply_to: None,
-        };
-        let err = channel.send(&binding, msg).expect_err("no bot");
-        assert!(
-            err.to_string().contains("telegram bot not initialized"),
-            "{err}"
-        );
-    }
-
-    // ─── Topic orphan fix tests (Sprint 14 PR-AI) ─────────────────────
-
-    #[test]
-    fn lookup_topic_for_instance_finds_existing() {
-        let home = tmp_home("lookup_existing");
-        register_topic(&home, 42, "alice");
-        register_topic(&home, 99, "bob");
-        assert_eq!(lookup_topic_for_instance(&home, "alice"), Some(42));
-        assert_eq!(lookup_topic_for_instance(&home, "bob"), Some(99));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn lookup_topic_for_instance_returns_none_when_missing() {
-        let home = tmp_home("lookup_missing");
-        register_topic(&home, 42, "alice");
-        assert_eq!(lookup_topic_for_instance(&home, "nonexistent"), None);
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    #[test]
-    fn lookup_topic_for_instance_returns_none_when_no_file() {
-        let home = PathBuf::from("/tmp/agend-test-no-topics-json-12345");
-        assert_eq!(lookup_topic_for_instance(&home, "any"), None);
-    }
-
-    /// create_topic_for_instance must reuse an existing topic from
-    /// topics.json instead of creating a new one. Without a live bot
-    /// the API call would fail, so the test seeds topics.json and
-    /// verifies the early return path.
-    #[test]
-    fn create_topic_for_instance_reuses_existing_topic() {
-        let home = tmp_home("create_reuse");
-        register_topic(&home, 77, "reuse-agent");
-        // No fleet.yaml / no bot token → if it tried to create, it would
-        // return None. But the lookup-before-create path should return 77.
-        let result = create_topic_for_instance(&home, "reuse-agent");
-        assert_eq!(
-            result,
-            Some(77),
-            "must reuse existing topic, not create new"
-        );
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// When no existing topic exists and no channel is configured,
-    /// create_topic_for_instance returns None (no bot to call).
-    #[test]
-    fn create_topic_for_instance_returns_none_without_config() {
-        let home = tmp_home("create_no_config");
-        let result = create_topic_for_instance(&home, "new-agent");
-        assert!(result.is_none(), "no config → no topic creation");
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// Invariant: `handle_message` must not call `block_on` — it runs inside
-    /// the telegram polling thread's tokio runtime, so nested `block_on`
-    /// panics. This grep-based test catches regressions.
-    #[test]
-    fn handle_message_body_has_no_block_on() {
-        let src = include_str!("inbound.rs");
-        // Extract handle_message body (from "async fn handle_message" to next top-level fn)
-        let start = src
-            .find("async fn handle_message(")
-            .expect("handle_message must exist");
-        // Find the next top-level function after handle_message
-        let rest = &src[start + 30..];
-        let end = rest
-            .find("\nfn ")
-            .or_else(|| rest.find("\nasync fn "))
-            .or_else(|| rest.find("\npub fn "))
-            .or_else(|| rest.find("\npub(crate) fn "))
-            .unwrap_or(rest.len());
-        let body = &rest[..end];
-        assert!(
-            !body.contains("block_on"),
-            "handle_message must not call block_on (nested runtime panic). Found block_on in body."
-        );
-    }
-
-    // ─── Sprint 23 P1: Topic-cache validate-before-reuse tests ────────
-
-    /// `invalidate_and_recreate_topic` must strip the stale topic from
-    /// `topics.json` so `create_topic_for_instance` doesn't short-circuit
-    /// on the dead id. Without a live bot the creation step returns `None`,
-    /// but the invalidation side-effect is the load-bearing contract.
-    #[test]
-    fn invalidate_and_recreate_strips_stale_topic_from_registry() {
-        let home = tmp_home("invalidate-recreate-strip");
-        register_topic(&home, 42, "agent-x");
-        register_topic(&home, 99, "agent-y");
-
-        // No fleet.yaml / no bot → create_topic_for_instance returns None,
-        // but the stale entry must still be gone from the registry.
-        let result = invalidate_and_recreate_topic(&home, "agent-x", 42);
-        assert!(result.is_none(), "no bot → creation fails");
-
-        let reg = load_topic_registry(&home);
-        assert!(
-            !reg.contains_key(&42),
-            "stale topic 42 must be unregistered"
-        );
-        // Unrelated entry survives.
-        assert_eq!(reg.get(&99), Some(&"agent-y".to_string()));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// `invalidate_and_recreate_topic` must clear `topic_id` in fleet.yaml
-    /// so the stale value doesn't persist across daemon restarts.
-    #[test]
-    fn invalidate_and_recreate_clears_fleet_yaml_topic_id() {
-        let home = tmp_home("invalidate-recreate-yaml");
-        let yaml = "\
-instances:
-  agent-x:
-    command: /bin/true
-    topic_id: 42
-  agent-y:
-    command: /bin/true
-    topic_id: 99
-";
-        std::fs::write(home.join("fleet.yaml"), yaml).expect("write fleet.yaml");
-        register_topic(&home, 42, "agent-x");
-
-        let _ = invalidate_and_recreate_topic(&home, "agent-x", 42);
-
-        // fleet.yaml must no longer have topic_id: 42 for agent-x.
-        let config =
-            crate::fleet::FleetConfig::load(&home.join("fleet.yaml")).expect("load fleet.yaml");
-        let agent_x = config.instances.get("agent-x").expect("agent-x exists");
-        assert_eq!(
-            agent_x.topic_id, None,
-            "topic_id must be cleared after invalidation"
-        );
-        // agent-y untouched.
-        let agent_y = config.instances.get("agent-y").expect("agent-y exists");
-        assert_eq!(agent_y.topic_id, Some(99));
-        std::fs::remove_dir_all(&home).ok();
-    }
-
-    /// `invalidate_and_recreate_topic` must NOT delete the instance from
-    /// fleet.yaml — only the topic_id field is cleared. This is the key
-    /// difference from `cleanup_deleted_topic` which tears down the
-    /// entire instance.
-    #[test]
-    fn invalidate_and_recreate_preserves_instance_in_fleet_yaml() {
-        let home = tmp_home("invalidate-preserves-instance");
-        let yaml = "\
-instances:
-  agent-x:
-    command: /bin/true
-    topic_id: 42
-    role: important
-";
-        std::fs::write(home.join("fleet.yaml"), yaml).expect("write fleet.yaml");
-        register_topic(&home, 42, "agent-x");
-
-        let _ = invalidate_and_recreate_topic(&home, "agent-x", 42);
-
-        let fleet_yaml = std::fs::read_to_string(home.join("fleet.yaml")).expect("read fleet.yaml");
-        assert!(
-            fleet_yaml.contains("agent-x"),
-            "instance must survive invalidation; yaml:\n{fleet_yaml}"
-        );
-        assert!(
-            fleet_yaml.contains("important"),
-            "instance role must survive; yaml:\n{fleet_yaml}"
-        );
-        std::fs::remove_dir_all(&home).ok();
     }
 
     /// `try_telegram_reply_from` must retry with a recreated topic when
